+++ src/main/java/com/example/carsharing/service/impl/CarServiceImpl.java
@@ -14,7 +14,7 @@ public class CarServiceImpl implements CarService {
     private final CarRepository carRepository;
 
     @Override
-    public Car getById(Integer id) {
+    public Car getById(Long id) {
         return carRepository.findById(id).orElseThrow(() ->
                 new NoSuchElementException("Can't find car by id " + id));
     }
@@ -25,7 +25,7 @@ public class CarServiceImpl implements CarService {
     }
 
     @Override
-    public void deleteById(Integer id) {
+    public void deleteById(Long id) {
         carRepository.deleteById(id);
     }
